Tests/LibWeb/Text/expected/HTML/image-srcset-future-compat-height-descriptor.txt

currentSrc: 120.png
